fix(frontend/pages): moves pages into protected and fix includes
