+++ frontend/app/(protected)/Pages/Create/components/reactionButton.tsx
@@ -1,7 +1,7 @@
 import { Button, Stack, YStack, Text, XStack } from 'tamagui';
 import { Link } from 'expo-router';
-import { useNavigationData } from '../../../context/navigationContext';
-import { Reaction } from "../../../context/appletContext";
+import { useNavigationData } from '../../../../context/navigationContext';
+import { Reaction } from "../../../../context/appletContext";
 
 
 export function ReactionButton({ index, reaction } : { index : number, reaction: Reaction }) {
@@ -14,7 +14,7 @@ export function ReactionButton({ index, reaction } : { index : number, reaction:
         setNavigationData({
           currentService: reaction.service ? reaction.service : "",
           actionType: !reaction.service ? "reaction" : "modify",
-          id: reaction.id ? reaction.id : ""
+          reactionId: reaction.id ? reaction.id : ""
         })}}
       borderWidth="$1"
       borderColor="$color"
