+++ frontend/app/(protected)/Pages/Create/form.tsx
@@ -1,7 +1,7 @@
 import { Button, ScrollView, Stack, YStack, Text, View, XStack, Square, H2 } from 'tamagui'
-import { useNavigationData } from '../../context/navigationContext'
+import { useNavigationData } from '../../../context/navigationContext'
 import { Link } from 'expo-router'
-import { useApplet } from '../../context/appletContext';
+import { useApplet } from '../../../context/appletContext';
 
 
 export default function ServicesScreen() {
@@ -14,17 +14,7 @@ export default function ServicesScreen() {
   const { navigationData } = useNavigationData();
 
   return (
-    <ScrollView
-    contentContainerStyle={{
-      flexGrow: 1, 
-      backgroundColor: '$background'
-    }}
-    style={{ flex: 1 }}>
-      <YStack paddingVertical="$4" width="100%" alignItems='center' gap="$2" >
-        {actions.map((action, i) => [
-            
-        ])}
-      </YStack>
+    <ScrollView>
     </ScrollView>
   )
 }
